// components/infinite-canvas/types/canvas-plugin.ts
export type CanvasNodeToolbarItem = {
  id: string;
  title: string;
  label: string;
  icon: ReactNode;
  onClick: () => void;
  active?: boolean;
  danger?: boolean;
};